backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
python-dotenv==1.0.1
pypdf==4.3.1
chromadb==0.5.5
sentence-transformers==3.0.1
numpy==1.26.4
openai==1.51.2
httpx==0.27.2
